# Add in memcpy only blit for opaque blits