@MyreMylar 2022-08-31
A big old table of results covering different bit depths. General rule that seems to hold is that if the width of the blit goes above  32ish then this PR starts to perform much better. Below that width it is a bit more of a competition. The SSE code appears particularly good with odd widths in small sizes.

I shared a better formatted version of this table [here, on google sheets.](https://docs.google.com/spreadsheets/d/1Biln4jrIM1U1x86IO0UdQLzIp_d9_FofdPU4-ACLfV8/edit?usp=sharing)

Blit size | Iterations | Bit depth | Without Mecpy PR | With Memcpy PR | % Change
-- | -- | -- | -- | -- | -- |
0,1 | 50000 | 8 | 12.805 | 13.012 | -1.59%
1, 1 | 50000 | 8 | 15.014 | 14.013 | 7.14%
1, 2 | 50000 | 8 | 15.013 | 14.013 | 7.14%
2, 2 | 50000 | 8 | 15.014 | 15.013 | 0.01%
3, 4 | 50000 | 8 | 15.013 | 15.014 | -0.01%
4, 4 | 50000 | 8 | 15.014 | 15.014 | 0.00%
7, 8 | 50000 | 8 | 17.931 | 17.015 | 5.38%
8, 8 | 50000 | 8 | 15.014 | 15.013 | 0.01%
15, 16 | 50000 | 8 | 17.015 | 16.015 | 6.24%
16, 16 | 50000 | 8 | 17.015 | 17.015 | 0.00%
31, 32 | 10000 | 8 | 6.005 | 6.004 | 0.02%
32, 32 | 10000 | 8 | 4.003 | 4.004 | -0.02%
63, 64 | 10000 | 8 | 6.006 | 6.025 | -0.32%
64, 64 | 10000 | 8 | 10.009 | 6.006 | 66.65%
127, 128 | 10000 | 8 | 10.009 | 10.009 | 0.00%
128, 128 | 10000 | 8 | 18.016 | 9.989 | 80.36%
255, 256 | 10000 | 8 | 31.028 | 21.019 | 47.62%
256, 256 | 10000 | 8 | 44.041 | 20.019 | 120.00%
511, 512 | 10000 | 8 | 132.119 | 82.075 | 60.97%
512, 512 | 10000 | 8 | 147.134 | 85.078 | 72.94%
1023, 1024 | 10000 | 8 | 447.86 | 273.671 | 63.65%
1024, 1024 | 10000 | 8 | 578.527 | 260.23 | 122.31%
0,1 | 50000 | 16 | 13.012 | 14.013 | -7.14%
1, 1 | 50000 | 16 | 13.012 | 14.013 | -7.14%
1, 2 | 50000 | 16 | 14.012 | 14.013 | -0.01%
2, 2 | 50000 | 16 | 15.013 | 14.013 | 7.14%
3, 4 | 50000 | 16 | 15.014 | 16.015 | -6.25%
4, 4 | 50000 | 16 | 15.013 | 15.014 | -0.01%
7, 8 | 50000 | 16 | 15.014 | 15.014 | 0.00%
8, 8 | 50000 | 16 | 15.014 | 15.013 | 0.01%
15, 16 | 50000 | 16 | 17.016 | 17.016 | 0.00%
16, 16 | 50000 | 16 | 17.016 | 17.015 | 0.01%
31, 32 | 10000 | 16 | 4.004 | 5.006 | -20.02%
32, 32 | 10000 | 16 | 7.005 | 5.004 | 39.99%
63, 64 | 10000 | 16 | 7.006 | 6.005 | 16.67%
64, 64 | 10000 | 16 | 11.01 | 6.005 | 83.35%
127, 128 | 10000 | 16 | 16.014 | 12.011 | 33.33%
128, 128 | 10000 | 16 | 23.022 | 11.01 | 109.10%
255, 256 | 10000 | 16 | 58.052 | 33.031 | 75.75%
256, 256 | 10000 | 16 | 72.065 | 30.028 | 139.99%
511, 512 | 10000 | 16 | 245.222 | 154.143 | 59.09%
512, 512 | 10000 | 16 | 297.692 | 152.138 | 95.67%
1023, 1024 | 10000 | 16 | 877.798 | 535.213 | 64.01%
1024, 1024 | 10000 | 16 | 1156.051 | 498.674 | 131.82%
0,1 | 50000 | 24 | 11.921 | 13.012 | -8.38%
1, 1 | 50000 | 24 | 13.021 | 13.012 | 0.07%
1, 2 | 50000 | 24 | 14.004 | 14.013 | -0.06%
2, 2 | 50000 | 24 | 14.013 | 15.013 | -6.66%
3, 4 | 50000 | 24 | 14.013 | 14.013 | 0.00%
4, 4 | 50000 | 24 | 13.012 | 14.013 | -7.14%
7, 8 | 50000 | 24 | 15.023 | 15.013 | 0.07%
8, 8 | 50000 | 24 | 14.921 | 15.014 | -0.62%
15, 16 | 50000 | 24 | 18.016 | 18.016 | 0.00%
16, 16 | 50000 | 24 | 17.016 | 18.016 | -5.55%
31, 32 | 10000 | 24 | 5.005 | 4.004 | 25.00%
32, 32 | 10000 | 24 | 4.004 | 5.004 | -19.98%
63, 64 | 10000 | 24 | 8.006 | 7.007 | 14.26%
64, 64 | 10000 | 24 | 10.009 | 7.006 | 42.86%
127, 128 | 10000 | 24 | 23.02 | 14.013 | 64.28%
128, 128 | 10000 | 24 | 29.028 | 14.013 | 107.15%
255, 256 | 10000 | 24 | 95.086 | 57.053 | 66.66%
256, 256 | 10000 | 24 | 108.098 | 55.049 | 96.37%
511, 512 | 10000 | 24 | 351.318 | 223.527 | 57.17%
512, 512 | 10000 | 24 | 446.407 | 220.204 | 102.72%
1023, 1024 | 10000 | 24 | 1269.973 | 769.142 | 65.12%
1024, 1024 | 10000 | 24 | 1700.735 | 744.066 | 128.57%
0,1 | 50000 | 32 | 13.012 | 14.013 | -7.14%
1, 1 | 50000 | 32 | 14.013 | 13.011 | 7.70%
1, 2 | 50000 | 32 | 14.013 | 15.014 | -6.67%
2, 2 | 50000 | 32 | 14.013 | 14.013 | 0.00%
3, 4 | 50000 | 32 | 15.013 | 14.013 | 7.14%
4, 4 | 50000 | 32 | 15.014 | 14.013 | 7.14%
7, 8 | 50000 | 32 | 16.014 | 15.014 | 6.66%
8, 8 | 50000 | 32 | 16.015 | 15.014 | 6.67%
15, 16 | 50000 | 32 | 17.015 | 18.016 | -5.56%
16, 16 | 50000 | 32 | 22.021 | 19.017 | 15.80%
31, 32 | 10000 | 32 | 4.003 | 4.004 | -0.02%
32, 32 | 10000 | 32 | 8.006 | 6.006 | 33.30%
63, 64 | 10000 | 32 | 9.008 | 7.007 | 28.56%
64, 64 | 10000 | 32 | 13.012 | 7.005 | 85.75%
127, 128 | 10000 | 32 | 28.025 | 19.018 | 47.36%
128, 128 | 10000 | 32 | 40.037 | 16.015 | 150.00%
255, 256 | 10000 | 32 | 141.128 | 80.073 | 76.25%
256, 256 | 10000 | 32 | 144.131 | 78.07 | 84.62%
511, 512 | 10000 | 32 | 450.225 | 280.583 | 60.46%
512, 512 | 10000 | 32 | 592.762 | 274.255 | 116.14%
1023, 1024 | 10000 | 32 | 1694.539 | 1005.177 | 68.58%
1024, 1024 | 10000 | 32 | 2249.045 | 936.968 | 140.03%